+++ yeti/genomics/roitools.py
@@ -1900,7 +1900,7 @@ class SegmentChain(object):
             Line from a `BED`_ file, containing 4 or more columns
 
         extra_columns: int or list optional
-            Extra, non-BED columns in `BED`_ file corresponding to feature
+            Extra, non-BED columns in :term:`BED X+Y`_ format file corresponding to feature
             attributes. This is common in `ENCODE`_-specific `BED`_ variants.
             
             if `extra-columns` is: